% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuteFileBrowser.R
\name{cuteFileBrowserUI}
\alias{cuteFileBrowserUI}
\alias{cuteFileBrowserServer}
\title{Cute File Browser}
\usage{
cuteFileBrowserUI(inputId)

cuteFileBrowserServer(inputId, rootDirectory,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{String. The input identifier used to access the value.}

\item{rootDirectory}{String. The begin directory to browser. Read-only access needed.}

\item{session}{Shiny default reactive domain.}
}
\description{
Create a materialize file input
}
\examples{
if (interactive()) {
library(shiny)
library(cuteFileBrowser)
ui <-
 fluidPage(
    theme = "bootstrap.css",
    titlePanel("Cute File Browser"),
    cuteFileBrowserUI(
      inputId = "file"
    )
 )

server <- function(input, output, session) {
  cuteFileBrowserServer(
    inputId = "file",
    rootDirectory = "~"
  )

  observeEvent(input$file, {
    showNotification(input$file)
  });
}

shinyApp(ui = ui, server = server)
}

}
